requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.1.1
idna==3.7
mysqlclient==2.2.4
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
